+++ UnitsNet.Serialization.JsonNet/UnitsNetBaseJsonConverter.cs
@@ -25,7 +25,7 @@ namespace UnitsNet.Serialization.JsonNet
 
         /// <summary>
         /// Register custom types so that the converter can instantiate these quantities.
-        /// Instead of calling <see cref="Quantity.From(QuantityValue,UnitKey)"/>, the <see cref="Activator"/> will be used to instantiate the object.
+        /// Instead of calling <see cref="Quantity.From(double,UnitKey)"/>, the <see cref="Activator"/> will be used to instantiate the object.
         /// It is therefore assumed that the constructor of <paramref name="quantity"/> is specified with <c>new T(double value, typeof(<paramref name="unit"/>) unit)</c>.
         /// Registering the same <paramref name="unit"/> multiple times, it will overwrite the one registered.
         /// </summary>
@@ -112,7 +112,7 @@ namespace UnitsNet.Serialization.JsonNet
                     $"Cannot create instance of type {registeredQuantity.Name} from value {valueUnit.Value} and unit {valueUnit.Unit}.");
             }
 
-            return Quantity.From(QuantityValue.FromDoubleRounded(valueUnit.Value), unit);
+            return Quantity.From((double)QuantityValue.FromDoubleRounded(valueUnit.Value), unit);
         }
 
         private (Type? Quantity, Type? Unit) GetRegisteredType(string unit)
